Fixes issue #2004 - Refactor core/jpms into core/api and core/impl
