+++ core/api/src/main/java/cloud/piranha/core/api/ModuleLayerProcessor.java
@@ -4,14 +4,14 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *   1. Redistributions of source code must retain the above copyright notice,
- *      this list of conditions and the following disclaimer.
- *   2. Redistributions in binary form must reproduce the above copyright
- *      notice, this list of conditions and the following disclaimer in the
- *      documentation and/or other materials provided with the distribution.
- *   3. Neither the name of the copyright holder nor the names of its
- *      contributors may be used to endorse or promote products derived from
- *      this software without specific prior written permission.
+ *  1. Redistributions of source code must retain the above copyright notice,
+ *     this list of conditions and the following disclaimer.
+ *  2. Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in the
+ *     documentation and/or other materials provided with the distribution.
+ *  3. Neither the name of the copyright holder nor the names of its
+ *     contributors may be used to endorse or promote products derived from
+ *     this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
@@ -24,19 +24,19 @@
  * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
+ *
  */
+package cloud.piranha.core.api;
 
 /**
- * The Piranha Modular - JPMS integration module.
- *
- * <p>
- *  This module delivers the default implementation of the {@link java.lang.module} API.
- * </p>
- *
+ * The module layer processor
  * @author Thiago Henrique Hupner
  */
-module cloud.piranha.core.jpms {
+public interface ModuleLayerProcessor {
 
-    exports cloud.piranha.core.jpms;
-    requires cloud.piranha.resource.api;
+    /**
+     * Process the module layer according to the properties
+     * @param controller the controller
+     */
+    void processModuleLayerOptions(ModuleLayer.Controller controller);
 }
