+++ core/impl/src/main/java/cloud/piranha/core/impl/DefaultModuleReader.java
@@ -27,7 +27,7 @@
  *
  */
 
-package cloud.piranha.core.jpms;
+package cloud.piranha.core.impl;
 
 import cloud.piranha.resource.api.Resource;
