+++ core/pom.xml
@@ -25,7 +25,6 @@
     <modules>
         <module>api</module>
         <module>impl</module>
-        <module>jpms</module>
         <module>servlet-api</module>
     </modules>
 </project>
