+++ micro/distribution/pom.xml
@@ -17,7 +17,7 @@
     <dependencies>
         <dependency>
             <groupId>cloud.piranha.core</groupId>
-            <artifactId>piranha-core-jpms</artifactId>
+            <artifactId>piranha-core-impl</artifactId>
             <version>${project.version}</version>
             <optional>true</optional>
         </dependency>
