+++ micro/loader/pom.xml
@@ -18,7 +18,7 @@
         <!-- compile -->
         <dependency>
             <groupId>cloud.piranha.core</groupId>
-            <artifactId>piranha-core-jpms</artifactId>
+            <artifactId>piranha-core-impl</artifactId>
             <version>${project.version}</version>
             <optional>true</optional>
         </dependency>
