+++ server/pom.xml
@@ -37,12 +37,6 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>cloud.piranha.core</groupId>
-            <artifactId>piranha-core-jpms</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>cloud.piranha.extension</groupId>
             <artifactId>piranha-extension-security-file</artifactId>
             <version>${project.version}</version>
